--- src/ktt/lib/httpserver/handler/Authenticator.java
@@ -1,4 +1,0 @@
-package ktt.lib.httpserver.handler;
-
-public interface Authenticator {
-}
