+++ src/ktt/lib/httpserver/handler_depreciated/Authenticator.java
@@ -1,0 +1,4 @@
+package ktt.lib.httpserver.handler_depreciated;
+
+public interface Authenticator {
+}
